population_data_analysis/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[tool.poetry]
name = "population_data_analysis"
version = "0.0.0"
description = "Module only for API contracts between services."
authors = [ "Matthew Daw <mattdaw7@gmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  cachetools = "*"
  matplotlib = "*"
  mlflow = "*"
  pandas = "*"
  pydantic-settings = "^2.2.1"
  python-dotenv = "^0.21.0"
  python = "^3.12.0"
  seaborn = "*"
  snowflake-connector-python = "*"
  statsmodels = "*"
  structlog = "^24.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
